update showcase client
