--- test/showcase-echo-client/showcase/.eslintignore
@@ -1,7 +1,0 @@
-**/node_modules
-**/.coverage
-build/
-docs/
-protos/
-system-test/
-samples/generated/
