--- test/showcase-echo-client/showcase/.eslintrc.json
@@ -1,3 +1,0 @@
-{
-  "extends": "./node_modules/gts"
-}
