--- test/showcase-echo-client/showcase/.gitignore
@@ -1,14 +1,0 @@
-**/*.log
-**/node_modules
-/.coverage
-/coverage
-/.nyc_output
-/docs/
-/out/
-/build/
-system-test/secrets.js
-system-test/*key.json
-*.lock
-.DS_Store
-package-lock.json
-__pycache__
