--- test/showcase-echo-client/showcase/package.json
@@ -1,59 +1,0 @@
-{
-  "name": "showcase",
-  "version": "0.1.0",
-  "description": "Showcase client for Node.js",
-  "repository": "googleapis/nodejs-showcase",
-  "license": "Apache-2.0",
-  "author": "Google LLC",
-  "main": "build/src/index.js",
-  "files": [
-    "build/src",
-    "build/protos"
-  ],
-  "keywords": [
-    "google apis client",
-    "google api client",
-    "google apis",
-    "google api",
-    "google",
-    "google cloud platform",
-    "google cloud",
-    "cloud",
-    "google showcase",
-    "showcase",
-    "echo",
-    "sequence service"
-  ],
-  "scripts": {
-    "clean": "gts clean",
-    "compile": "tsc -p . && cp -r protos build/ && minifyProtoJson",
-    "compile-protos": "compileProtos src",
-    "docs": "jsdoc -c .jsdoc.js",
-    "fix": "gts fix",
-    "lint": "gts check",
-    "prepare": "npm run compile-protos && npm run compile",
-    "system-test": "c8 mocha build/system-test",
-    "test": "c8 mocha build/test"
-  },
-  "dependencies": {
-    "google-gax": "^4.0.3"
-  },
-  "devDependencies": {
-    "@types/mocha": "^10.0.1",
-    "@types/node": "^18.11.18",
-    "@types/sinon": "^10.0.16",
-    "c8": "^8.0.1",
-    "gapic-tools": "^0.1.8",
-    "gts": "5.0.0",
-    "jsdoc": "^4.0.2",
-    "jsdoc-fresh": "^2.0.2",
-    "jsdoc-region-tag": "^2.0.1",
-    "mocha": "^10.2.0",
-    "pack-n-play": "^1.0.0-2",
-    "sinon": "^15.2.0",
-    "typescript": "5.1.6"
-  },
-  "engines": {
-    "node": ">=v14"
-  }
-}
