--- test/showcase-echo-client/showcase/src/index.ts
@@ -1,27 +1,0 @@
-// Copyright 2023 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// ** This file is automatically generated by gapic-generator-typescript. **
-// ** https://github.com/googleapis/gapic-generator-typescript **
-// ** All changes to this file may be overwritten. **
-
-import * as v1beta1 from './v1beta1';
-const EchoClient = v1beta1.EchoClient;
-type EchoClient = v1beta1.EchoClient;
-const SequenceServiceClient = v1beta1.SequenceServiceClient;
-type SequenceServiceClient = v1beta1.SequenceServiceClient;
-export {v1beta1, EchoClient, SequenceServiceClient};
-export default {v1beta1, EchoClient, SequenceServiceClient};
-import * as protos from '../protos/protos';
-export {protos}
