--- test/showcase-echo-client/showcase/src/v1beta1/echo_client_config.json
@@ -1,62 +1,0 @@
-{
-  "interfaces": {
-    "google.showcase.v1beta1.Echo": {
-      "retry_codes": {
-        "non_idempotent": [],
-        "idempotent": [
-          "DEADLINE_EXCEEDED",
-          "UNAVAILABLE"
-        ]
-      },
-      "retry_params": {
-        "default": {
-          "initial_retry_delay_millis": 100,
-          "retry_delay_multiplier": 1.3,
-          "max_retry_delay_millis": 60000,
-          "initial_rpc_timeout_millis": 60000,
-          "rpc_timeout_multiplier": 1,
-          "max_rpc_timeout_millis": 60000,
-          "total_timeout_millis": 600000
-        }
-      },
-      "methods": {
-        "Echo": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        },
-        "Expand": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        },
-        "Collect": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        },
-        "Chat": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        },
-        "PagedExpand": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        },
-        "PagedExpandLegacy": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        },
-        "PagedExpandLegacyMapped": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        },
-        "Wait": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        },
-        "Block": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        }
-      }
-    }
-  }
-}
