--- test/showcase-echo-client/showcase/src/v1beta1/echo_proto_list.json
@@ -1,4 +1,0 @@
-[
-  "../../protos/google/showcase/v1beta1/echo.proto",
-  "../../protos/google/showcase/v1beta1/sequence.proto"
-]
