--- test/showcase-echo-client/showcase/src/v1beta1/sequence_service_client_config.json
@@ -1,50 +1,0 @@
-{
-  "interfaces": {
-    "google.showcase.v1beta1.SequenceService": {
-      "retry_codes": {
-        "non_idempotent": [],
-        "idempotent": [
-          "DEADLINE_EXCEEDED",
-          "UNAVAILABLE"
-        ]
-      },
-      "retry_params": {
-        "default": {
-          "initial_retry_delay_millis": 100,
-          "retry_delay_multiplier": 1.3,
-          "max_retry_delay_millis": 60000,
-          "initial_rpc_timeout_millis": 60000,
-          "rpc_timeout_multiplier": 1,
-          "max_rpc_timeout_millis": 60000,
-          "total_timeout_millis": 600000
-        }
-      },
-      "methods": {
-        "CreateSequence": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        },
-        "CreateStreamingSequence": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        },
-        "GetSequenceReport": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        },
-        "GetStreamingSequenceReport": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        },
-        "AttemptSequence": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        },
-        "AttemptStreamingSequence": {
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
-        }
-      }
-    }
-  }
-}
