--- test/showcase-echo-client/showcase/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "extends": "./node_modules/gts/tsconfig-google.json",
-  "compilerOptions": {
-    "rootDir": ".",
-    "outDir": "build",
-    "resolveJsonModule": true,
-    "lib": [
-      "es2018",
-      "dom"
-    ]
-  },
-  "include": [
-    "src/*.ts",
-    "src/**/*.ts",
-    "test/*.ts",
-    "test/**/*.ts",
-    "system-test/*.ts"
-  ]
-}
