+++ test/showcase-echo-client/src/v1beta1/sequence_service_client.ts
@@ -15,19 +15,10 @@
 // ** This file is automatically generated by gapic-generator-typescript. **
 // ** https://github.com/googleapis/gapic-generator-typescript **
 // ** All changes to this file may be overwritten. **
-//@ts-nocheck
+
 /* global window */
 import type * as gax from 'google-gax';
-import type {
-  Callback,
-  CallOptions,
-  Descriptors,
-  ClientOptions,
-  IamClient,
-  IamProtos,
-  LocationsClient,
-  LocationProtos,
-} from 'google-gax';
+import type {Callback, CallOptions, Descriptors, ClientOptions, IamClient, IamProtos, LocationsClient, LocationProtos} from 'google-gax';
 import {PassThrough} from 'stream';
 import * as protos from '../../protos/protos';
 import jsonProtos = require('../../protos/protos.json');
@@ -105,22 +96,14 @@ export class SequenceServiceClient {
    *     const client = new SequenceServiceClient({fallback: 'rest'}, gax);
    *     ```
    */
-  constructor(
-    opts?: ClientOptions,
-    gaxInstance?: typeof gax | typeof gax.fallback
-  ) {
+  constructor(opts?: ClientOptions, gaxInstance?: typeof gax | typeof gax.fallback) {
     // Ensure that options include all the required fields.
     const staticMembers = this.constructor as typeof SequenceServiceClient;
-    const servicePath =
-      opts?.servicePath || opts?.apiEndpoint || staticMembers.servicePath;
-    this._providedCustomServicePath = !!(
-      opts?.servicePath || opts?.apiEndpoint
-    );
+    const servicePath = opts?.servicePath || opts?.apiEndpoint || staticMembers.servicePath;
+    this._providedCustomServicePath = !!(opts?.servicePath || opts?.apiEndpoint);
     const port = opts?.port || staticMembers.port;
     const clientConfig = opts?.clientConfig ?? {};
-    const fallback =
-      opts?.fallback ??
-      (typeof window !== 'undefined' && typeof window?.fetch === 'function');
+    const fallback = opts?.fallback ?? (typeof window !== 'undefined' && typeof window?.fetch === 'function');
     opts = Object.assign({servicePath, port, clientConfig, fallback}, opts);
 
     // If scopes are unset in options and we're connecting to a non-default endpoint, set scopes just in case.
@@ -143,7 +126,7 @@ export class SequenceServiceClient {
     this._opts = opts;
 
     // Save the auth object to the client, for use by other methods.
-    this.auth = this._gaxGrpc.auth as gax.GoogleAuth;
+    this.auth = (this._gaxGrpc.auth as gax.GoogleAuth);
 
     // Set useJWTAccessWithScope on the auth object.
     this.auth.useJWTAccessWithScope = true;
@@ -156,14 +139,18 @@ export class SequenceServiceClient {
       this.auth.defaultScopes = staticMembers.scopes;
     }
     this.iamClient = new this._gaxModule.IamClient(this._gaxGrpc, opts);
-
+  
     this.locationsClient = new this._gaxModule.LocationsClient(
       this._gaxGrpc,
       opts
     );
+  
 
     // Determine the client header string.
-    const clientHeader = [`gax/${this._gaxModule.version}`, `gapic/${version}`];
+    const clientHeader = [
+      `gax/${this._gaxModule.version}`,
+      `gapic/${version}`,
+    ];
     if (typeof process !== 'undefined' && 'versions' in process) {
       clientHeader.push(`gl-node/${process.versions.node}`);
     } else {
@@ -171,7 +158,7 @@ export class SequenceServiceClient {
     }
     if (!opts.fallback) {
       clientHeader.push(`grpc/${this._gaxGrpc.grpcVersion}`);
-    } else if (opts.fallback === 'rest') {
+    } else if (opts.fallback === 'rest' ) {
       clientHeader.push(`rest/${this._gaxGrpc.grpcVersion}`);
     }
     if (opts.libName && opts.libVersion) {
@@ -201,20 +188,13 @@ export class SequenceServiceClient {
     // Some of the methods on this service provide streaming responses.
     // Provide descriptors for these.
     this.descriptors.stream = {
-      attemptStreamingSequence: new this._gaxModule.StreamDescriptor(
-        this._gaxModule.StreamType.SERVER_STREAMING,
-        opts.fallback === 'rest',
-        this._opts.newRetry
-      ),
+      attemptStreamingSequence: new this._gaxModule.StreamDescriptor(this._gaxModule.StreamType.SERVER_STREAMING, opts.fallback === 'rest')
     };
 
     // Put together the default options sent with requests.
     this._defaults = this._gaxGrpc.constructSettings(
-      'google.showcase.v1beta1.SequenceService',
-      gapicConfig as gax.ClientConfig,
-      opts.clientConfig || {},
-      {'x-goog-api-client': clientHeader.join(' ')}
-    );
+        'google.showcase.v1beta1.SequenceService', gapicConfig as gax.ClientConfig,
+        opts.clientConfig || {}, {'x-goog-api-client': clientHeader.join(' ')});
 
     // Set up a dictionary of "inner API calls"; the core implementation
     // of calling the API is handled in `google-gax`, with this code
@@ -245,54 +225,39 @@ export class SequenceServiceClient {
     // Put together the "service stub" for
     // google.showcase.v1beta1.SequenceService.
     this.sequenceServiceStub = this._gaxGrpc.createStub(
-      this._opts.fallback
-        ? (this._protos as protobuf.Root).lookupService(
-            'google.showcase.v1beta1.SequenceService'
-          )
-        : // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        this._opts.fallback ?
+          (this._protos as protobuf.Root).lookupService('google.showcase.v1beta1.SequenceService') :
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
           (this._protos as any).google.showcase.v1beta1.SequenceService,
-      this._opts,
-      this._providedCustomServicePath
-    ) as Promise<{[method: string]: Function}>;
+        this._opts, this._providedCustomServicePath) as Promise<{[method: string]: Function}>;
 
     // Iterate over each of the methods that the service provides
     // and create an API call method for each.
-    const sequenceServiceStubMethods = [
-      'createSequence',
-      'createStreamingSequence',
-      'getSequenceReport',
-      'getStreamingSequenceReport',
-      'attemptSequence',
-      'attemptStreamingSequence',
-    ];
+    const sequenceServiceStubMethods =
+        ['createSequence', 'createStreamingSequence', 'getSequenceReport', 'getStreamingSequenceReport', 'attemptSequence', 'attemptStreamingSequence'];
     for (const methodName of sequenceServiceStubMethods) {
       const callPromise = this.sequenceServiceStub.then(
-        stub =>
-          (...args: Array<{}>) => {
-            if (this._terminated) {
-              if (methodName in this.descriptors.stream) {
-                const stream = new PassThrough();
-                setImmediate(() => {
-                  stream.emit(
-                    'error',
-                    new this._gaxModule.GoogleError(
-                      'The client has already been closed.'
-                    )
-                  );
-                });
-                return stream;
-              }
-              return Promise.reject('The client has already been closed.');
+        stub => (...args: Array<{}>) => {
+          if (this._terminated) {
+            if (methodName in this.descriptors.stream) {
+              const stream = new PassThrough();
+              setImmediate(() => {
+                stream.emit('error', new this._gaxModule.GoogleError('The client has already been closed.'));
+              });
+              return stream;
             }
-            const func = stub[methodName];
-            return func.apply(stub, args);
-          },
-        (err: Error | null | undefined) => () => {
+            return Promise.reject('The client has already been closed.');
+          }
+          const func = stub[methodName];
+          return func.apply(stub, args);
+        },
+        (err: Error|null|undefined) => () => {
           throw err;
-        }
-      );
+        });
 
-      const descriptor = this.descriptors.stream[methodName] || undefined;
+      const descriptor =
+        this.descriptors.stream[methodName] ||
+        undefined;
       const apiCall = this._gaxModule.createApiCall(
         callPromise,
         this._defaults[methodName],
@@ -346,9 +311,8 @@ export class SequenceServiceClient {
    * Return the project ID used by this class.
    * @returns {Promise} A promise that resolves to string containing the project ID.
    */
-  getProjectId(
-    callback?: Callback<string, undefined, undefined>
-  ): Promise<string> | void {
+  getProjectId(callback?: Callback<string, undefined, undefined>):
+      Promise<string>|void {
     if (callback) {
       this.auth.getProjectId(callback);
       return;
@@ -359,77 +323,62 @@ export class SequenceServiceClient {
   // -------------------
   // -- Service calls --
   // -------------------
-  /**
-   * Creates a sequence.
-   *
-   * @param {Object} request
-   *   The request object that will be sent.
-   * @param {google.showcase.v1beta1.Sequence} request.sequence
-   * @param {object} [options]
-   *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
-   * @returns {Promise} - The promise which resolves to an array.
-   *   The first element of the array is an object representing {@link protos.google.showcase.v1beta1.Sequence|Sequence}.
-   *   Please see the {@link https://github.com/googleapis/gax-nodejs/blob/master/client-libraries.md#regular-methods | documentation }
-   *   for more details and examples.
-   * @example <caption>include:samples/generated/v1beta1/sequence_service.create_sequence.js</caption>
-   * region_tag:localhost_v1beta1_generated_SequenceService_CreateSequence_async
-   */
+/**
+ * Creates a sequence.
+ *
+ * @param {Object} request
+ *   The request object that will be sent.
+ * @param {google.showcase.v1beta1.Sequence} request.sequence
+ * @param {object} [options]
+ *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
+ * @returns {Promise} - The promise which resolves to an array.
+ *   The first element of the array is an object representing {@link protos.google.showcase.v1beta1.Sequence|Sequence}.
+ *   Please see the {@link https://github.com/googleapis/gax-nodejs/blob/master/client-libraries.md#regular-methods | documentation }
+ *   for more details and examples.
+ * @example <caption>include:samples/generated/v1beta1/sequence_service.create_sequence.js</caption>
+ * region_tag:localhost_v1beta1_generated_SequenceService_CreateSequence_async
+ */
   createSequence(
-    request?: protos.google.showcase.v1beta1.ICreateSequenceRequest,
-    options?: CallOptions
-  ): Promise<
-    [
-      protos.google.showcase.v1beta1.ISequence,
-      protos.google.showcase.v1beta1.ICreateSequenceRequest | undefined,
-      {} | undefined
-    ]
-  >;
+      request?: protos.google.showcase.v1beta1.ICreateSequenceRequest,
+      options?: CallOptions):
+      Promise<[
+        protos.google.showcase.v1beta1.ISequence,
+        protos.google.showcase.v1beta1.ICreateSequenceRequest|undefined, {}|undefined
+      ]>;
   createSequence(
-    request: protos.google.showcase.v1beta1.ICreateSequenceRequest,
-    options: CallOptions,
-    callback: Callback<
-      protos.google.showcase.v1beta1.ISequence,
-      protos.google.showcase.v1beta1.ICreateSequenceRequest | null | undefined,
-      {} | null | undefined
-    >
-  ): void;
-  createSequence(
-    request: protos.google.showcase.v1beta1.ICreateSequenceRequest,
-    callback: Callback<
-      protos.google.showcase.v1beta1.ISequence,
-      protos.google.showcase.v1beta1.ICreateSequenceRequest | null | undefined,
-      {} | null | undefined
-    >
-  ): void;
-  createSequence(
-    request?: protos.google.showcase.v1beta1.ICreateSequenceRequest,
-    optionsOrCallback?:
-      | CallOptions
-      | Callback<
+      request: protos.google.showcase.v1beta1.ICreateSequenceRequest,
+      options: CallOptions,
+      callback: Callback<
           protos.google.showcase.v1beta1.ISequence,
-          | protos.google.showcase.v1beta1.ICreateSequenceRequest
-          | null
-          | undefined,
-          {} | null | undefined
-        >,
-    callback?: Callback<
-      protos.google.showcase.v1beta1.ISequence,
-      protos.google.showcase.v1beta1.ICreateSequenceRequest | null | undefined,
-      {} | null | undefined
-    >
-  ): Promise<
-    [
-      protos.google.showcase.v1beta1.ISequence,
-      protos.google.showcase.v1beta1.ICreateSequenceRequest | undefined,
-      {} | undefined
-    ]
-  > | void {
+          protos.google.showcase.v1beta1.ICreateSequenceRequest|null|undefined,
+          {}|null|undefined>): void;
+  createSequence(
+      request: protos.google.showcase.v1beta1.ICreateSequenceRequest,
+      callback: Callback<
+          protos.google.showcase.v1beta1.ISequence,
+          protos.google.showcase.v1beta1.ICreateSequenceRequest|null|undefined,
+          {}|null|undefined>): void;
+  createSequence(
+      request?: protos.google.showcase.v1beta1.ICreateSequenceRequest,
+      optionsOrCallback?: CallOptions|Callback<
+          protos.google.showcase.v1beta1.ISequence,
+          protos.google.showcase.v1beta1.ICreateSequenceRequest|null|undefined,
+          {}|null|undefined>,
+      callback?: Callback<
+          protos.google.showcase.v1beta1.ISequence,
+          protos.google.showcase.v1beta1.ICreateSequenceRequest|null|undefined,
+          {}|null|undefined>):
+      Promise<[
+        protos.google.showcase.v1beta1.ISequence,
+        protos.google.showcase.v1beta1.ICreateSequenceRequest|undefined, {}|undefined
+      ]>|void {
     request = request || {};
     let options: CallOptions;
     if (typeof optionsOrCallback === 'function' && callback === undefined) {
       callback = optionsOrCallback;
       options = {};
-    } else {
+    }
+    else {
       options = optionsOrCallback as CallOptions;
     }
     options = options || {};
@@ -438,429 +387,335 @@ export class SequenceServiceClient {
     this.initialize();
     return this.innerApiCalls.createSequence(request, options, callback);
   }
-  /**
-   * Creates a sequence.
-   *
-   * @param {Object} request
-   *   The request object that will be sent.
-   * @param {google.showcase.v1beta1.StreamingSequence} request.streamingSequence
-   * @param {object} [options]
-   *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
-   * @returns {Promise} - The promise which resolves to an array.
-   *   The first element of the array is an object representing {@link protos.google.showcase.v1beta1.StreamingSequence|StreamingSequence}.
-   *   Please see the {@link https://github.com/googleapis/gax-nodejs/blob/master/client-libraries.md#regular-methods | documentation }
-   *   for more details and examples.
-   * @example <caption>include:samples/generated/v1beta1/sequence_service.create_streaming_sequence.js</caption>
-   * region_tag:localhost_v1beta1_generated_SequenceService_CreateStreamingSequence_async
-   */
+/**
+ * Creates a sequence.
+ *
+ * @param {Object} request
+ *   The request object that will be sent.
+ * @param {google.showcase.v1beta1.StreamingSequence} request.streamingSequence
+ * @param {object} [options]
+ *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
+ * @returns {Promise} - The promise which resolves to an array.
+ *   The first element of the array is an object representing {@link protos.google.showcase.v1beta1.StreamingSequence|StreamingSequence}.
+ *   Please see the {@link https://github.com/googleapis/gax-nodejs/blob/master/client-libraries.md#regular-methods | documentation }
+ *   for more details and examples.
+ * @example <caption>include:samples/generated/v1beta1/sequence_service.create_streaming_sequence.js</caption>
+ * region_tag:localhost_v1beta1_generated_SequenceService_CreateStreamingSequence_async
+ */
   createStreamingSequence(
-    request?: protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest,
-    options?: CallOptions
-  ): Promise<
-    [
-      protos.google.showcase.v1beta1.IStreamingSequence,
-      (
-        | protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest
-        | undefined
-      ),
-      {} | undefined
-    ]
-  >;
+      request?: protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest,
+      options?: CallOptions):
+      Promise<[
+        protos.google.showcase.v1beta1.IStreamingSequence,
+        protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest|undefined, {}|undefined
+      ]>;
   createStreamingSequence(
-    request: protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest,
-    options: CallOptions,
-    callback: Callback<
-      protos.google.showcase.v1beta1.IStreamingSequence,
-      | protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest
-      | null
-      | undefined,
-      {} | null | undefined
-    >
-  ): void;
-  createStreamingSequence(
-    request: protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest,
-    callback: Callback<
-      protos.google.showcase.v1beta1.IStreamingSequence,
-      | protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest
-      | null
-      | undefined,
-      {} | null | undefined
-    >
-  ): void;
-  createStreamingSequence(
-    request?: protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest,
-    optionsOrCallback?:
-      | CallOptions
-      | Callback<
+      request: protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest,
+      options: CallOptions,
+      callback: Callback<
           protos.google.showcase.v1beta1.IStreamingSequence,
-          | protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest
-          | null
-          | undefined,
-          {} | null | undefined
-        >,
-    callback?: Callback<
-      protos.google.showcase.v1beta1.IStreamingSequence,
-      | protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest
-      | null
-      | undefined,
-      {} | null | undefined
-    >
-  ): Promise<
-    [
-      protos.google.showcase.v1beta1.IStreamingSequence,
-      (
-        | protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest
-        | undefined
-      ),
-      {} | undefined
-    ]
-  > | void {
+          protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest|null|undefined,
+          {}|null|undefined>): void;
+  createStreamingSequence(
+      request: protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest,
+      callback: Callback<
+          protos.google.showcase.v1beta1.IStreamingSequence,
+          protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest|null|undefined,
+          {}|null|undefined>): void;
+  createStreamingSequence(
+      request?: protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest,
+      optionsOrCallback?: CallOptions|Callback<
+          protos.google.showcase.v1beta1.IStreamingSequence,
+          protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest|null|undefined,
+          {}|null|undefined>,
+      callback?: Callback<
+          protos.google.showcase.v1beta1.IStreamingSequence,
+          protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest|null|undefined,
+          {}|null|undefined>):
+      Promise<[
+        protos.google.showcase.v1beta1.IStreamingSequence,
+        protos.google.showcase.v1beta1.ICreateStreamingSequenceRequest|undefined, {}|undefined
+      ]>|void {
     request = request || {};
     let options: CallOptions;
     if (typeof optionsOrCallback === 'function' && callback === undefined) {
       callback = optionsOrCallback;
       options = {};
-    } else {
+    }
+    else {
       options = optionsOrCallback as CallOptions;
     }
     options = options || {};
     options.otherArgs = options.otherArgs || {};
     options.otherArgs.headers = options.otherArgs.headers || {};
     this.initialize();
-    return this.innerApiCalls.createStreamingSequence(
-      request,
-      options,
-      callback
-    );
+    return this.innerApiCalls.createStreamingSequence(request, options, callback);
   }
-  /**
-   * Retrieves a sequence.
-   *
-   * @param {Object} request
-   *   The request object that will be sent.
-   * @param {string} request.name
-   * @param {object} [options]
-   *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
-   * @returns {Promise} - The promise which resolves to an array.
-   *   The first element of the array is an object representing {@link protos.google.showcase.v1beta1.SequenceReport|SequenceReport}.
-   *   Please see the {@link https://github.com/googleapis/gax-nodejs/blob/master/client-libraries.md#regular-methods | documentation }
-   *   for more details and examples.
-   * @example <caption>include:samples/generated/v1beta1/sequence_service.get_sequence_report.js</caption>
-   * region_tag:localhost_v1beta1_generated_SequenceService_GetSequenceReport_async
-   */
+/**
+ * Retrieves a sequence.
+ *
+ * @param {Object} request
+ *   The request object that will be sent.
+ * @param {string} request.name
+ * @param {object} [options]
+ *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
+ * @returns {Promise} - The promise which resolves to an array.
+ *   The first element of the array is an object representing {@link protos.google.showcase.v1beta1.SequenceReport|SequenceReport}.
+ *   Please see the {@link https://github.com/googleapis/gax-nodejs/blob/master/client-libraries.md#regular-methods | documentation }
+ *   for more details and examples.
+ * @example <caption>include:samples/generated/v1beta1/sequence_service.get_sequence_report.js</caption>
+ * region_tag:localhost_v1beta1_generated_SequenceService_GetSequenceReport_async
+ */
   getSequenceReport(
-    request?: protos.google.showcase.v1beta1.IGetSequenceReportRequest,
-    options?: CallOptions
-  ): Promise<
-    [
-      protos.google.showcase.v1beta1.ISequenceReport,
-      protos.google.showcase.v1beta1.IGetSequenceReportRequest | undefined,
-      {} | undefined
-    ]
-  >;
+      request?: protos.google.showcase.v1beta1.IGetSequenceReportRequest,
+      options?: CallOptions):
+      Promise<[
+        protos.google.showcase.v1beta1.ISequenceReport,
+        protos.google.showcase.v1beta1.IGetSequenceReportRequest|undefined, {}|undefined
+      ]>;
   getSequenceReport(
-    request: protos.google.showcase.v1beta1.IGetSequenceReportRequest,
-    options: CallOptions,
-    callback: Callback<
-      protos.google.showcase.v1beta1.ISequenceReport,
-      | protos.google.showcase.v1beta1.IGetSequenceReportRequest
-      | null
-      | undefined,
-      {} | null | undefined
-    >
-  ): void;
-  getSequenceReport(
-    request: protos.google.showcase.v1beta1.IGetSequenceReportRequest,
-    callback: Callback<
-      protos.google.showcase.v1beta1.ISequenceReport,
-      | protos.google.showcase.v1beta1.IGetSequenceReportRequest
-      | null
-      | undefined,
-      {} | null | undefined
-    >
-  ): void;
-  getSequenceReport(
-    request?: protos.google.showcase.v1beta1.IGetSequenceReportRequest,
-    optionsOrCallback?:
-      | CallOptions
-      | Callback<
+      request: protos.google.showcase.v1beta1.IGetSequenceReportRequest,
+      options: CallOptions,
+      callback: Callback<
           protos.google.showcase.v1beta1.ISequenceReport,
-          | protos.google.showcase.v1beta1.IGetSequenceReportRequest
-          | null
-          | undefined,
-          {} | null | undefined
-        >,
-    callback?: Callback<
-      protos.google.showcase.v1beta1.ISequenceReport,
-      | protos.google.showcase.v1beta1.IGetSequenceReportRequest
-      | null
-      | undefined,
-      {} | null | undefined
-    >
-  ): Promise<
-    [
-      protos.google.showcase.v1beta1.ISequenceReport,
-      protos.google.showcase.v1beta1.IGetSequenceReportRequest | undefined,
-      {} | undefined
-    ]
-  > | void {
+          protos.google.showcase.v1beta1.IGetSequenceReportRequest|null|undefined,
+          {}|null|undefined>): void;
+  getSequenceReport(
+      request: protos.google.showcase.v1beta1.IGetSequenceReportRequest,
+      callback: Callback<
+          protos.google.showcase.v1beta1.ISequenceReport,
+          protos.google.showcase.v1beta1.IGetSequenceReportRequest|null|undefined,
+          {}|null|undefined>): void;
+  getSequenceReport(
+      request?: protos.google.showcase.v1beta1.IGetSequenceReportRequest,
+      optionsOrCallback?: CallOptions|Callback<
+          protos.google.showcase.v1beta1.ISequenceReport,
+          protos.google.showcase.v1beta1.IGetSequenceReportRequest|null|undefined,
+          {}|null|undefined>,
+      callback?: Callback<
+          protos.google.showcase.v1beta1.ISequenceReport,
+          protos.google.showcase.v1beta1.IGetSequenceReportRequest|null|undefined,
+          {}|null|undefined>):
+      Promise<[
+        protos.google.showcase.v1beta1.ISequenceReport,
+        protos.google.showcase.v1beta1.IGetSequenceReportRequest|undefined, {}|undefined
+      ]>|void {
     request = request || {};
     let options: CallOptions;
     if (typeof optionsOrCallback === 'function' && callback === undefined) {
       callback = optionsOrCallback;
       options = {};
-    } else {
+    }
+    else {
       options = optionsOrCallback as CallOptions;
     }
     options = options || {};
     options.otherArgs = options.otherArgs || {};
     options.otherArgs.headers = options.otherArgs.headers || {};
-    options.otherArgs.headers['x-goog-request-params'] =
-      this._gaxModule.routingHeader.fromParams({
-        name: request.name ?? '',
-      });
+    options.otherArgs.headers[
+      'x-goog-request-params'
+    ] = this._gaxModule.routingHeader.fromParams({
+      'name': request.name ?? '',
+    });
     this.initialize();
     return this.innerApiCalls.getSequenceReport(request, options, callback);
   }
-  /**
-   * Retrieves a sequence.
-   *
-   * @param {Object} request
-   *   The request object that will be sent.
-   * @param {string} request.name
-   * @param {object} [options]
-   *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
-   * @returns {Promise} - The promise which resolves to an array.
-   *   The first element of the array is an object representing {@link protos.google.showcase.v1beta1.StreamingSequenceReport|StreamingSequenceReport}.
-   *   Please see the {@link https://github.com/googleapis/gax-nodejs/blob/master/client-libraries.md#regular-methods | documentation }
-   *   for more details and examples.
-   * @example <caption>include:samples/generated/v1beta1/sequence_service.get_streaming_sequence_report.js</caption>
-   * region_tag:localhost_v1beta1_generated_SequenceService_GetStreamingSequenceReport_async
-   */
+/**
+ * Retrieves a sequence.
+ *
+ * @param {Object} request
+ *   The request object that will be sent.
+ * @param {string} request.name
+ * @param {object} [options]
+ *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
+ * @returns {Promise} - The promise which resolves to an array.
+ *   The first element of the array is an object representing {@link protos.google.showcase.v1beta1.StreamingSequenceReport|StreamingSequenceReport}.
+ *   Please see the {@link https://github.com/googleapis/gax-nodejs/blob/master/client-libraries.md#regular-methods | documentation }
+ *   for more details and examples.
+ * @example <caption>include:samples/generated/v1beta1/sequence_service.get_streaming_sequence_report.js</caption>
+ * region_tag:localhost_v1beta1_generated_SequenceService_GetStreamingSequenceReport_async
+ */
   getStreamingSequenceReport(
-    request?: protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest,
-    options?: CallOptions
-  ): Promise<
-    [
-      protos.google.showcase.v1beta1.IStreamingSequenceReport,
-      (
-        | protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest
-        | undefined
-      ),
-      {} | undefined
-    ]
-  >;
+      request?: protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest,
+      options?: CallOptions):
+      Promise<[
+        protos.google.showcase.v1beta1.IStreamingSequenceReport,
+        protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest|undefined, {}|undefined
+      ]>;
   getStreamingSequenceReport(
-    request: protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest,
-    options: CallOptions,
-    callback: Callback<
-      protos.google.showcase.v1beta1.IStreamingSequenceReport,
-      | protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest
-      | null
-      | undefined,
-      {} | null | undefined
-    >
-  ): void;
-  getStreamingSequenceReport(
-    request: protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest,
-    callback: Callback<
-      protos.google.showcase.v1beta1.IStreamingSequenceReport,
-      | protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest
-      | null
-      | undefined,
-      {} | null | undefined
-    >
-  ): void;
-  getStreamingSequenceReport(
-    request?: protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest,
-    optionsOrCallback?:
-      | CallOptions
-      | Callback<
+      request: protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest,
+      options: CallOptions,
+      callback: Callback<
           protos.google.showcase.v1beta1.IStreamingSequenceReport,
-          | protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest
-          | null
-          | undefined,
-          {} | null | undefined
-        >,
-    callback?: Callback<
-      protos.google.showcase.v1beta1.IStreamingSequenceReport,
-      | protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest
-      | null
-      | undefined,
-      {} | null | undefined
-    >
-  ): Promise<
-    [
-      protos.google.showcase.v1beta1.IStreamingSequenceReport,
-      (
-        | protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest
-        | undefined
-      ),
-      {} | undefined
-    ]
-  > | void {
+          protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest|null|undefined,
+          {}|null|undefined>): void;
+  getStreamingSequenceReport(
+      request: protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest,
+      callback: Callback<
+          protos.google.showcase.v1beta1.IStreamingSequenceReport,
+          protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest|null|undefined,
+          {}|null|undefined>): void;
+  getStreamingSequenceReport(
+      request?: protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest,
+      optionsOrCallback?: CallOptions|Callback<
+          protos.google.showcase.v1beta1.IStreamingSequenceReport,
+          protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest|null|undefined,
+          {}|null|undefined>,
+      callback?: Callback<
+          protos.google.showcase.v1beta1.IStreamingSequenceReport,
+          protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest|null|undefined,
+          {}|null|undefined>):
+      Promise<[
+        protos.google.showcase.v1beta1.IStreamingSequenceReport,
+        protos.google.showcase.v1beta1.IGetStreamingSequenceReportRequest|undefined, {}|undefined
+      ]>|void {
     request = request || {};
     let options: CallOptions;
     if (typeof optionsOrCallback === 'function' && callback === undefined) {
       callback = optionsOrCallback;
       options = {};
-    } else {
+    }
+    else {
       options = optionsOrCallback as CallOptions;
     }
     options = options || {};
     options.otherArgs = options.otherArgs || {};
     options.otherArgs.headers = options.otherArgs.headers || {};
-    options.otherArgs.headers['x-goog-request-params'] =
-      this._gaxModule.routingHeader.fromParams({
-        name: request.name ?? '',
-      });
+    options.otherArgs.headers[
+      'x-goog-request-params'
+    ] = this._gaxModule.routingHeader.fromParams({
+      'name': request.name ?? '',
+    });
     this.initialize();
-    return this.innerApiCalls.getStreamingSequenceReport(
-      request,
-      options,
-      callback
-    );
+    return this.innerApiCalls.getStreamingSequenceReport(request, options, callback);
   }
-  /**
-   * Attempts a sequence.
-   *
-   * @param {Object} request
-   *   The request object that will be sent.
-   * @param {string} request.name
-   * @param {object} [options]
-   *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
-   * @returns {Promise} - The promise which resolves to an array.
-   *   The first element of the array is an object representing {@link protos.google.protobuf.Empty|Empty}.
-   *   Please see the {@link https://github.com/googleapis/gax-nodejs/blob/master/client-libraries.md#regular-methods | documentation }
-   *   for more details and examples.
-   * @example <caption>include:samples/generated/v1beta1/sequence_service.attempt_sequence.js</caption>
-   * region_tag:localhost_v1beta1_generated_SequenceService_AttemptSequence_async
-   */
+/**
+ * Attempts a sequence.
+ *
+ * @param {Object} request
+ *   The request object that will be sent.
+ * @param {string} request.name
+ * @param {object} [options]
+ *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
+ * @returns {Promise} - The promise which resolves to an array.
+ *   The first element of the array is an object representing {@link protos.google.protobuf.Empty|Empty}.
+ *   Please see the {@link https://github.com/googleapis/gax-nodejs/blob/master/client-libraries.md#regular-methods | documentation }
+ *   for more details and examples.
+ * @example <caption>include:samples/generated/v1beta1/sequence_service.attempt_sequence.js</caption>
+ * region_tag:localhost_v1beta1_generated_SequenceService_AttemptSequence_async
+ */
   attemptSequence(
-    request?: protos.google.showcase.v1beta1.IAttemptSequenceRequest,
-    options?: CallOptions
-  ): Promise<
-    [
-      protos.google.protobuf.IEmpty,
-      protos.google.showcase.v1beta1.IAttemptSequenceRequest | undefined,
-      {} | undefined
-    ]
-  >;
+      request?: protos.google.showcase.v1beta1.IAttemptSequenceRequest,
+      options?: CallOptions):
+      Promise<[
+        protos.google.protobuf.IEmpty,
+        protos.google.showcase.v1beta1.IAttemptSequenceRequest|undefined, {}|undefined
+      ]>;
   attemptSequence(
-    request: protos.google.showcase.v1beta1.IAttemptSequenceRequest,
-    options: CallOptions,
-    callback: Callback<
-      protos.google.protobuf.IEmpty,
-      protos.google.showcase.v1beta1.IAttemptSequenceRequest | null | undefined,
-      {} | null | undefined
-    >
-  ): void;
-  attemptSequence(
-    request: protos.google.showcase.v1beta1.IAttemptSequenceRequest,
-    callback: Callback<
-      protos.google.protobuf.IEmpty,
-      protos.google.showcase.v1beta1.IAttemptSequenceRequest | null | undefined,
-      {} | null | undefined
-    >
-  ): void;
-  attemptSequence(
-    request?: protos.google.showcase.v1beta1.IAttemptSequenceRequest,
-    optionsOrCallback?:
-      | CallOptions
-      | Callback<
+      request: protos.google.showcase.v1beta1.IAttemptSequenceRequest,
+      options: CallOptions,
+      callback: Callback<
           protos.google.protobuf.IEmpty,
-          | protos.google.showcase.v1beta1.IAttemptSequenceRequest
-          | null
-          | undefined,
-          {} | null | undefined
-        >,
-    callback?: Callback<
-      protos.google.protobuf.IEmpty,
-      protos.google.showcase.v1beta1.IAttemptSequenceRequest | null | undefined,
-      {} | null | undefined
-    >
-  ): Promise<
-    [
-      protos.google.protobuf.IEmpty,
-      protos.google.showcase.v1beta1.IAttemptSequenceRequest | undefined,
-      {} | undefined
-    ]
-  > | void {
+          protos.google.showcase.v1beta1.IAttemptSequenceRequest|null|undefined,
+          {}|null|undefined>): void;
+  attemptSequence(
+      request: protos.google.showcase.v1beta1.IAttemptSequenceRequest,
+      callback: Callback<
+          protos.google.protobuf.IEmpty,
+          protos.google.showcase.v1beta1.IAttemptSequenceRequest|null|undefined,
+          {}|null|undefined>): void;
+  attemptSequence(
+      request?: protos.google.showcase.v1beta1.IAttemptSequenceRequest,
+      optionsOrCallback?: CallOptions|Callback<
+          protos.google.protobuf.IEmpty,
+          protos.google.showcase.v1beta1.IAttemptSequenceRequest|null|undefined,
+          {}|null|undefined>,
+      callback?: Callback<
+          protos.google.protobuf.IEmpty,
+          protos.google.showcase.v1beta1.IAttemptSequenceRequest|null|undefined,
+          {}|null|undefined>):
+      Promise<[
+        protos.google.protobuf.IEmpty,
+        protos.google.showcase.v1beta1.IAttemptSequenceRequest|undefined, {}|undefined
+      ]>|void {
     request = request || {};
     let options: CallOptions;
     if (typeof optionsOrCallback === 'function' && callback === undefined) {
       callback = optionsOrCallback;
       options = {};
-    } else {
+    }
+    else {
       options = optionsOrCallback as CallOptions;
     }
     options = options || {};
     options.otherArgs = options.otherArgs || {};
     options.otherArgs.headers = options.otherArgs.headers || {};
-    options.otherArgs.headers['x-goog-request-params'] =
-      this._gaxModule.routingHeader.fromParams({
-        name: request.name ?? '',
-      });
+    options.otherArgs.headers[
+      'x-goog-request-params'
+    ] = this._gaxModule.routingHeader.fromParams({
+      'name': request.name ?? '',
+    });
     this.initialize();
     return this.innerApiCalls.attemptSequence(request, options, callback);
   }
 
-  /**
-   * Attempts a streaming sequence.
-   *
-   * @param {Object} request
-   *   The request object that will be sent.
-   * @param {string} request.name
-   * @param {object} [options]
-   *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
-   * @returns {Stream}
-   *   An object stream which emits {@link protos.google.showcase.v1beta1.AttemptStreamingSequenceResponse|AttemptStreamingSequenceResponse} on 'data' event.
-   *   Please see the {@link https://github.com/googleapis/gax-nodejs/blob/master/client-libraries.md#server-streaming | documentation }
-   *   for more details and examples.
-   * @example <caption>include:samples/generated/v1beta1/sequence_service.attempt_streaming_sequence.js</caption>
-   * region_tag:localhost_v1beta1_generated_SequenceService_AttemptStreamingSequence_async
-   */
+/**
+ * Attempts a streaming sequence.
+ *
+ * @param {Object} request
+ *   The request object that will be sent.
+ * @param {string} request.name
+ * @param {object} [options]
+ *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
+ * @returns {Stream}
+ *   An object stream which emits {@link protos.google.showcase.v1beta1.AttemptStreamingSequenceResponse|AttemptStreamingSequenceResponse} on 'data' event.
+ *   Please see the {@link https://github.com/googleapis/gax-nodejs/blob/master/client-libraries.md#server-streaming | documentation }
+ *   for more details and examples.
+ * @example <caption>include:samples/generated/v1beta1/sequence_service.attempt_streaming_sequence.js</caption>
+ * region_tag:localhost_v1beta1_generated_SequenceService_AttemptStreamingSequence_async
+ */
   attemptStreamingSequence(
-    request?: protos.google.showcase.v1beta1.IAttemptStreamingSequenceRequest,
-    options?: CallOptions
-  ): gax.CancellableStream {
+      request?: protos.google.showcase.v1beta1.IAttemptStreamingSequenceRequest,
+      options?: CallOptions):
+    gax.CancellableStream{
     request = request || {};
     options = options || {};
     options.otherArgs = options.otherArgs || {};
     options.otherArgs.headers = options.otherArgs.headers || {};
-    options.otherArgs.headers['x-goog-request-params'] =
-      this._gaxModule.routingHeader.fromParams({
-        name: request.name ?? '',
-      });
+    options.otherArgs.headers[
+      'x-goog-request-params'
+    ] = this._gaxModule.routingHeader.fromParams({
+      'name': request.name ?? '',
+    });
     this.initialize();
     return this.innerApiCalls.attemptStreamingSequence(request, options);
   }
 
-  /**
-   * Gets the access control policy for a resource. Returns an empty policy
-   * if the resource exists and does not have a policy set.
-   *
-   * @param {Object} request
-   *   The request object that will be sent.
-   * @param {string} request.resource
-   *   REQUIRED: The resource for which the policy is being requested.
-   *   See the operation documentation for the appropriate value for this field.
-   * @param {Object} [request.options]
-   *   OPTIONAL: A `GetPolicyOptions` object for specifying options to
-   *   `GetIamPolicy`. This field is only used by Cloud IAM.
-   *
-   *   This object should have the same structure as {@link google.iam.v1.GetPolicyOptions | GetPolicyOptions}.
-   * @param {Object} [options]
-   *   Optional parameters. You can override the default settings for this call, e.g, timeout,
-   *   retries, paginations, etc. See {@link https://googleapis.github.io/gax-nodejs/interfaces/CallOptions.html | gax.CallOptions} for the details.
-   * @param {function(?Error, ?Object)} [callback]
-   *   The function which will be called with the result of the API call.
-   *
-   *   The second parameter to the callback is an object representing {@link google.iam.v1.Policy | Policy}.
-   * @returns {Promise} - The promise which resolves to an array.
-   *   The first element of the array is an object representing {@link google.iam.v1.Policy | Policy}.
-   *   The promise has a method named "cancel" which cancels the ongoing API call.
-   */
+/**
+ * Gets the access control policy for a resource. Returns an empty policy
+ * if the resource exists and does not have a policy set.
+ *
+ * @param {Object} request
+ *   The request object that will be sent.
+ * @param {string} request.resource
+ *   REQUIRED: The resource for which the policy is being requested.
+ *   See the operation documentation for the appropriate value for this field.
+ * @param {Object} [request.options]
+ *   OPTIONAL: A `GetPolicyOptions` object for specifying options to
+ *   `GetIamPolicy`. This field is only used by Cloud IAM.
+ *
+ *   This object should have the same structure as {@link google.iam.v1.GetPolicyOptions | GetPolicyOptions}.
+ * @param {Object} [options]
+ *   Optional parameters. You can override the default settings for this call, e.g, timeout,
+ *   retries, paginations, etc. See {@link https://googleapis.github.io/gax-nodejs/interfaces/CallOptions.html | gax.CallOptions} for the details.
+ * @param {function(?Error, ?Object)} [callback]
+ *   The function which will be called with the result of the API call.
+ *
+ *   The second parameter to the callback is an object representing {@link google.iam.v1.Policy | Policy}.
+ * @returns {Promise} - The promise which resolves to an array.
+ *   The first element of the array is an object representing {@link google.iam.v1.Policy | Policy}.
+ *   The promise has a method named "cancel" which cancels the ongoing API call.
+ */
   getIamPolicy(
     request: IamProtos.google.iam.v1.GetIamPolicyRequest,
     options?:
@@ -875,39 +730,39 @@ export class SequenceServiceClient {
       IamProtos.google.iam.v1.GetIamPolicyRequest | null | undefined,
       {} | null | undefined
     >
-  ): Promise<IamProtos.google.iam.v1.Policy> {
+  ):Promise<[IamProtos.google.iam.v1.Policy]> {
     return this.iamClient.getIamPolicy(request, options, callback);
   }
 
-  /**
-   * Returns permissions that a caller has on the specified resource. If the
-   * resource does not exist, this will return an empty set of
-   * permissions, not a NOT_FOUND error.
-   *
-   * Note: This operation is designed to be used for building
-   * permission-aware UIs and command-line tools, not for authorization
-   * checking. This operation may "fail open" without warning.
-   *
-   * @param {Object} request
-   *   The request object that will be sent.
-   * @param {string} request.resource
-   *   REQUIRED: The resource for which the policy detail is being requested.
-   *   See the operation documentation for the appropriate value for this field.
-   * @param {string[]} request.permissions
-   *   The set of permissions to check for the `resource`. Permissions with
-   *   wildcards (such as '*' or 'storage.*') are not allowed. For more
-   *   information see {@link https://cloud.google.com/iam/docs/overview#permissions | IAM Overview }.
-   * @param {Object} [options]
-   *   Optional parameters. You can override the default settings for this call, e.g, timeout,
-   *   retries, paginations, etc. See {@link https://googleapis.github.io/gax-nodejs/interfaces/CallOptions.html | gax.CallOptions} for the details.
-   * @param {function(?Error, ?Object)} [callback]
-   *   The function which will be called with the result of the API call.
-   *
-   *   The second parameter to the callback is an object representing {@link google.iam.v1.TestIamPermissionsResponse | TestIamPermissionsResponse}.
-   * @returns {Promise} - The promise which resolves to an array.
-   *   The first element of the array is an object representing {@link google.iam.v1.TestIamPermissionsResponse | TestIamPermissionsResponse}.
-   *   The promise has a method named "cancel" which cancels the ongoing API call.
-   */
+/**
+ * Returns permissions that a caller has on the specified resource. If the
+ * resource does not exist, this will return an empty set of
+ * permissions, not a NOT_FOUND error.
+ *
+ * Note: This operation is designed to be used for building
+ * permission-aware UIs and command-line tools, not for authorization
+ * checking. This operation may "fail open" without warning.
+ *
+ * @param {Object} request
+ *   The request object that will be sent.
+ * @param {string} request.resource
+ *   REQUIRED: The resource for which the policy detail is being requested.
+ *   See the operation documentation for the appropriate value for this field.
+ * @param {string[]} request.permissions
+ *   The set of permissions to check for the `resource`. Permissions with
+ *   wildcards (such as '*' or 'storage.*') are not allowed. For more
+ *   information see {@link https://cloud.google.com/iam/docs/overview#permissions | IAM Overview }.
+ * @param {Object} [options]
+ *   Optional parameters. You can override the default settings for this call, e.g, timeout,
+ *   retries, paginations, etc. See {@link https://googleapis.github.io/gax-nodejs/interfaces/CallOptions.html | gax.CallOptions} for the details.
+ * @param {function(?Error, ?Object)} [callback]
+ *   The function which will be called with the result of the API call.
+ *
+ *   The second parameter to the callback is an object representing {@link google.iam.v1.TestIamPermissionsResponse | TestIamPermissionsResponse}.
+ * @returns {Promise} - The promise which resolves to an array.
+ *   The first element of the array is an object representing {@link google.iam.v1.TestIamPermissionsResponse | TestIamPermissionsResponse}.
+ *   The promise has a method named "cancel" which cancels the ongoing API call.
+ */
   setIamPolicy(
     request: IamProtos.google.iam.v1.SetIamPolicyRequest,
     options?:
@@ -922,40 +777,40 @@ export class SequenceServiceClient {
       IamProtos.google.iam.v1.SetIamPolicyRequest | null | undefined,
       {} | null | undefined
     >
-  ): Promise<IamProtos.google.iam.v1.Policy> {
+  ):Promise<[IamProtos.google.iam.v1.Policy]> {
     return this.iamClient.setIamPolicy(request, options, callback);
   }
 
-  /**
-   * Returns permissions that a caller has on the specified resource. If the
-   * resource does not exist, this will return an empty set of
-   * permissions, not a NOT_FOUND error.
-   *
-   * Note: This operation is designed to be used for building
-   * permission-aware UIs and command-line tools, not for authorization
-   * checking. This operation may "fail open" without warning.
-   *
-   * @param {Object} request
-   *   The request object that will be sent.
-   * @param {string} request.resource
-   *   REQUIRED: The resource for which the policy detail is being requested.
-   *   See the operation documentation for the appropriate value for this field.
-   * @param {string[]} request.permissions
-   *   The set of permissions to check for the `resource`. Permissions with
-   *   wildcards (such as '*' or 'storage.*') are not allowed. For more
-   *   information see {@link https://cloud.google.com/iam/docs/overview#permissions | IAM Overview }.
-   * @param {Object} [options]
-   *   Optional parameters. You can override the default settings for this call, e.g, timeout,
-   *   retries, paginations, etc. See {@link https://googleapis.github.io/gax-nodejs/interfaces/CallOptions.html | gax.CallOptions} for the details.
-   * @param {function(?Error, ?Object)} [callback]
-   *   The function which will be called with the result of the API call.
-   *
-   *   The second parameter to the callback is an object representing {@link google.iam.v1.TestIamPermissionsResponse | TestIamPermissionsResponse}.
-   * @returns {Promise} - The promise which resolves to an array.
-   *   The first element of the array is an object representing {@link google.iam.v1.TestIamPermissionsResponse | TestIamPermissionsResponse}.
-   *   The promise has a method named "cancel" which cancels the ongoing API call.
-   *
-   */
+/**
+ * Returns permissions that a caller has on the specified resource. If the
+ * resource does not exist, this will return an empty set of
+ * permissions, not a NOT_FOUND error.
+ *
+ * Note: This operation is designed to be used for building
+ * permission-aware UIs and command-line tools, not for authorization
+ * checking. This operation may "fail open" without warning.
+ *
+ * @param {Object} request
+ *   The request object that will be sent.
+ * @param {string} request.resource
+ *   REQUIRED: The resource for which the policy detail is being requested.
+ *   See the operation documentation for the appropriate value for this field.
+ * @param {string[]} request.permissions
+ *   The set of permissions to check for the `resource`. Permissions with
+ *   wildcards (such as '*' or 'storage.*') are not allowed. For more
+ *   information see {@link https://cloud.google.com/iam/docs/overview#permissions | IAM Overview }.
+ * @param {Object} [options]
+ *   Optional parameters. You can override the default settings for this call, e.g, timeout,
+ *   retries, paginations, etc. See {@link https://googleapis.github.io/gax-nodejs/interfaces/CallOptions.html | gax.CallOptions} for the details.
+ * @param {function(?Error, ?Object)} [callback]
+ *   The function which will be called with the result of the API call.
+ *
+ *   The second parameter to the callback is an object representing {@link google.iam.v1.TestIamPermissionsResponse | TestIamPermissionsResponse}.
+ * @returns {Promise} - The promise which resolves to an array.
+ *   The first element of the array is an object representing {@link google.iam.v1.TestIamPermissionsResponse | TestIamPermissionsResponse}.
+ *   The promise has a method named "cancel" which cancels the ongoing API call.
+ *
+ */
   testIamPermissions(
     request: IamProtos.google.iam.v1.TestIamPermissionsRequest,
     options?:
@@ -970,11 +825,11 @@ export class SequenceServiceClient {
       IamProtos.google.iam.v1.TestIamPermissionsRequest | null | undefined,
       {} | null | undefined
     >
-  ): Promise<IamProtos.google.iam.v1.TestIamPermissionsResponse> {
+  ):Promise<[IamProtos.google.iam.v1.TestIamPermissionsResponse]> {
     return this.iamClient.testIamPermissions(request, options, callback);
   }
 
-  /**
+/**
    * Gets information about a location.
    *
    * @param {Object} request
@@ -1014,7 +869,7 @@ export class SequenceServiceClient {
     return this.locationsClient.getLocation(request, options, callback);
   }
 
-  /**
+/**
    * Lists information about the supported locations for this service. Returns an iterable object.
    *
    * `for`-`await`-`of` syntax is used with the iterable to get response elements on-demand.
@@ -1062,7 +917,7 @@ export class SequenceServiceClient {
    * @param {string} sequence
    * @returns {string} Resource name string.
    */
-  sequencePath(sequence: string) {
+  sequencePath(sequence:string) {
     return this.pathTemplates.sequencePathTemplate.render({
       sequence: sequence,
     });
@@ -1085,7 +940,7 @@ export class SequenceServiceClient {
    * @param {string} sequence
    * @returns {string} Resource name string.
    */
-  sequenceReportPath(sequence: string) {
+  sequenceReportPath(sequence:string) {
     return this.pathTemplates.sequenceReportPathTemplate.render({
       sequence: sequence,
     });
@@ -1099,9 +954,7 @@ export class SequenceServiceClient {
    * @returns {string} A string representing the sequence.
    */
   matchSequenceFromSequenceReportName(sequenceReportName: string) {
-    return this.pathTemplates.sequenceReportPathTemplate.match(
-      sequenceReportName
-    ).sequence;
+    return this.pathTemplates.sequenceReportPathTemplate.match(sequenceReportName).sequence;
   }
 
   /**
@@ -1110,7 +963,7 @@ export class SequenceServiceClient {
    * @param {string} streaming_sequence
    * @returns {string} Resource name string.
    */
-  streamingSequencePath(streamingSequence: string) {
+  streamingSequencePath(streamingSequence:string) {
     return this.pathTemplates.streamingSequencePathTemplate.render({
       streaming_sequence: streamingSequence,
     });
@@ -1123,12 +976,8 @@ export class SequenceServiceClient {
    *   A fully-qualified path representing StreamingSequence resource.
    * @returns {string} A string representing the streaming_sequence.
    */
-  matchStreamingSequenceFromStreamingSequenceName(
-    streamingSequenceName: string
-  ) {
-    return this.pathTemplates.streamingSequencePathTemplate.match(
-      streamingSequenceName
-    ).streaming_sequence;
+  matchStreamingSequenceFromStreamingSequenceName(streamingSequenceName: string) {
+    return this.pathTemplates.streamingSequencePathTemplate.match(streamingSequenceName).streaming_sequence;
   }
 
   /**
@@ -1137,7 +986,7 @@ export class SequenceServiceClient {
    * @param {string} streaming_sequence
    * @returns {string} Resource name string.
    */
-  streamingSequenceReportPath(streamingSequence: string) {
+  streamingSequenceReportPath(streamingSequence:string) {
     return this.pathTemplates.streamingSequenceReportPathTemplate.render({
       streaming_sequence: streamingSequence,
     });
@@ -1150,12 +999,8 @@ export class SequenceServiceClient {
    *   A fully-qualified path representing StreamingSequenceReport resource.
    * @returns {string} A string representing the streaming_sequence.
    */
-  matchStreamingSequenceFromStreamingSequenceReportName(
-    streamingSequenceReportName: string
-  ) {
-    return this.pathTemplates.streamingSequenceReportPathTemplate.match(
-      streamingSequenceReportName
-    ).streaming_sequence;
+  matchStreamingSequenceFromStreamingSequenceReportName(streamingSequenceReportName: string) {
+    return this.pathTemplates.streamingSequenceReportPathTemplate.match(streamingSequenceReportName).streaming_sequence;
   }
 
   /**
